requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
pillow-heif>=0.13.0
pyinstaller>=6.0.0
